refactor: move withParticle to lib for error copy (#175)
